# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CommerceFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filtering, search, sorting
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, price_asc, price_desc, rating]
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /products/{productId}/reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Wishlist
...
  /wishlist/{productId}:
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove product from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # Users (admin)
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Analytics (wow endpoints)
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, admin]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    UserList:
      type: object
      required: [users, total, page, totalPages]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [customer, admin]

    # Product
    Product:
      type: object
      required: [id, name, description, price, stock, categoryId, imageUrl, rating, reviewCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        stock:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        createdAt:
          type: string
    ProductList:
      type: object
      required: [products, total, page, totalPages]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    ProductInput:
      type: object
      required: [name, description, price, stock, categoryId, imageUrl]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        stock:
          type: integer
        categoryId:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        stock:
          type: integer
        categoryId:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean

    # Category
...
    Review:
      type: object
      required: [id, userId, productId, rating, comment, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        productId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [rating, comment]
      properties:
        rating:
          type: integer
        comment:
          type: string

    # Wishlist
    WishlistItem:
      type: object
      required: [id, userId, productId, product, addedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        addedAt:
          type: string

    # Analytics
...