# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CommerceFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  stock: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  imageUrl: string;
  images?: string[];
  rating: number;
  reviewCount: number;
  isFeatured?: boolean;
  createdAt: string;
}