# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CommerceFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  stock: number;
  categoryId: number;
  imageUrl: string;
  images?: string[];
  isFeatured?: boolean;
}